@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.4.20'
+    ext.kotlin_version = '1.6.0'
 
     repositories {
         mavenCentral()
@@ -10,19 +10,20 @@ buildscript {
     }
 }
 
-apply plugin: 'kotlin'
-apply plugin: 'application'
+plugins {
+    id 'org.jetbrains.kotlin.jvm' version '1.6.0'
+    id 'application'
+}
 
 mainClassName = project.hasProperty("year") ? "v" + getProperty("year") + "/MainKt" : "INPUT-YEAR"
 
 repositories {
-    jcenter()
+    mavenCentral()
 }
 
 dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    testCompile 'junit:junit:4.12'
-    testCompile "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
+    testImplementation 'junit:junit:4.12'
+    implementation "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
 }
 
 test {
